refactor: rename script to utils
Also split up the UID command into sub-commands, so UID acts as a
namespace for `generate` and other commands we might need, e.g. `parse`.

Use `reporter.print` for output to stdout so the results of the tool can
be piped properly:

    d2 utils uid gen --json > uids.json
    d2 utils uid g 1 | cat

We might need to use something like
https://github.com/SheetJS/node-exit-on-epipe to make advanced pipe
commands work.